+++ tests/test_nice.py
@@ -214,6 +214,11 @@ class TestServantSpecial:
         assert response.status_code == 200
         assert response.json()["name"] == "The Holy Night's Aurora"
 
+    def test_JP_MC_English_name(self) -> None:
+        response = client.get("/nice/JP/MC/60?lang=en")
+        assert response.status_code == 200
+        assert response.json()["name"] == "Royal Brand"
+
     def test_empty_cv_illustrator_name(self) -> None:
         response = client.get("/nice/JP/svt/9941330?lore=true")
         assert response.status_code == 200